refactor(textarea): split textarea.rs into focused modules
Break up the 2426-line textarea.rs god object into smaller,
focused modules:

- cursor.rs: CursorPos, Cursor, CursorSet types
- selection.rs: Selection handling
- find_replace.rs: FindReplaceMode, FindOptions, FindMatch, FindReplaceState
- edit.rs: EditOperation for undo/redo
- mod.rs: Main TextArea struct and implementations

Removed inline tests in preparation for moving to tests/ directory.

Closes #198 (partial)
