+++ src/widget/textarea/mod.rs
@@ -9,6 +9,17 @@
 //! - Word wrap
 //! - Scrolling
 
+mod cursor;
+mod edit;
+mod find_replace;
+mod selection;
+
+pub use cursor::{Cursor, CursorPos, CursorSet};
+pub use find_replace::{FindMatch, FindOptions, FindReplaceMode, FindReplaceState};
+pub use selection::Selection;
+
+use edit::EditOperation;
+
 use super::syntax::{Language, SyntaxHighlighter, SyntaxTheme};
 use super::traits::{RenderContext, View, WidgetProps};
 use crate::event::Key;
@@ -18,364 +29,6 @@ use crate::{impl_props_builders, impl_styled_view};
 
 /// Maximum undo history size
 const MAX_UNDO_HISTORY: usize = 100;
-
-/// Maximum number of cursors allowed
-const MAX_CURSORS: usize = 100;
-
-// ============================================================================
-// Cursor Types
-// ============================================================================
-
-/// A cursor position in the text (line, column)
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct CursorPos {
-    /// Line index (0-based)
-    pub line: usize,
-    /// Column index (0-based)
-    pub col: usize,
-}
-
-impl CursorPos {
-    /// Create a new cursor position
-    pub fn new(line: usize, col: usize) -> Self {
-        Self { line, col }
-    }
-}
-
-impl From<(usize, usize)> for CursorPos {
-    fn from((line, col): (usize, usize)) -> Self {
-        Self { line, col }
-    }
-}
-
-impl From<CursorPos> for (usize, usize) {
-    fn from(pos: CursorPos) -> Self {
-        (pos.line, pos.col)
-    }
-}
-
-/// A cursor with optional selection anchor
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Cursor {
-    /// Current position
-    pub pos: CursorPos,
-    /// Selection anchor (if selecting)
-    pub anchor: Option<CursorPos>,
-}
-
-impl Cursor {
-    /// Create a new cursor at position
-    pub fn new(pos: CursorPos) -> Self {
-        Self { pos, anchor: None }
-    }
-
-    /// Create a cursor with selection
-    pub fn with_selection(pos: CursorPos, anchor: CursorPos) -> Self {
-        Self {
-            pos,
-            anchor: Some(anchor),
-        }
-    }
-
-    /// Get the selection as a Selection struct if selecting
-    pub fn selection(&self) -> Option<Selection> {
-        self.anchor.map(|anchor| {
-            let start = if self.pos < anchor {
-                (self.pos.line, self.pos.col)
-            } else {
-                (anchor.line, anchor.col)
-            };
-            let end = if self.pos < anchor {
-                (anchor.line, anchor.col)
-            } else {
-                (self.pos.line, self.pos.col)
-            };
-            Selection::new(start, end)
-        })
-    }
-
-    /// Check if this cursor is selecting
-    pub fn is_selecting(&self) -> bool {
-        self.anchor.is_some()
-    }
-
-    /// Start selection at current position
-    pub fn start_selection(&mut self) {
-        self.anchor = Some(self.pos);
-    }
-
-    /// Clear selection
-    pub fn clear_selection(&mut self) {
-        self.anchor = None;
-    }
-}
-
-/// Collection of cursors (always has at least one - the primary cursor)
-#[derive(Clone, Debug)]
-pub struct CursorSet {
-    /// All cursors, primary is at index 0
-    cursors: Vec<Cursor>,
-}
-
-impl CursorSet {
-    /// Create a new cursor set with a single cursor at position
-    pub fn new(pos: CursorPos) -> Self {
-        Self {
-            cursors: vec![Cursor::new(pos)],
-        }
-    }
-
-    /// Get the primary cursor (immutable)
-    pub fn primary(&self) -> &Cursor {
-        &self.cursors[0]
-    }
-
-    /// Get the primary cursor (mutable)
-    pub fn primary_mut(&mut self) -> &mut Cursor {
-        &mut self.cursors[0]
-    }
-
-    /// Get all cursors
-    #[allow(dead_code)]
-    pub fn all(&self) -> &[Cursor] {
-        &self.cursors
-    }
-
-    /// Get all cursors (mutable)
-    #[allow(dead_code)]
-    pub fn all_mut(&mut self) -> &mut [Cursor] {
-        &mut self.cursors
-    }
-
-    /// Get the number of cursors
-    pub fn len(&self) -> usize {
-        self.cursors.len()
-    }
-
-    /// Check if there's only one cursor
-    #[allow(dead_code)]
-    pub fn is_single(&self) -> bool {
-        self.cursors.len() == 1
-    }
-
-    /// Add a cursor at position
-    pub fn add(&mut self, cursor: Cursor) {
-        if self.cursors.len() < MAX_CURSORS {
-            self.cursors.push(cursor);
-            self.normalize();
-        }
-    }
-
-    /// Add a cursor at position (convenience method)
-    pub fn add_at(&mut self, pos: CursorPos) {
-        self.add(Cursor::new(pos));
-    }
-
-    /// Clear all secondary cursors, keeping only the primary
-    pub fn clear_secondary(&mut self) {
-        self.cursors.truncate(1);
-    }
-
-    /// Set the primary cursor position
-    pub fn set_primary(&mut self, pos: CursorPos) {
-        self.cursors[0].pos = pos;
-    }
-
-    /// Iterate over all cursors
-    pub fn iter(&self) -> impl Iterator<Item = &Cursor> {
-        self.cursors.iter()
-    }
-
-    /// Iterate over all cursors (mutable)
-    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut Cursor> {
-        self.cursors.iter_mut()
-    }
-
-    /// Sort cursors by position and merge overlapping ones
-    fn normalize(&mut self) {
-        // Sort by position (line, then column)
-        self.cursors.sort_by(|a, b| a.pos.cmp(&b.pos));
-
-        // Remove duplicates (cursors at same position)
-        self.cursors.dedup_by(|a, b| a.pos == b.pos);
-
-        // Ensure we always have at least one cursor
-        if self.cursors.is_empty() {
-            self.cursors.push(Cursor::new(CursorPos::new(0, 0)));
-        }
-    }
-
-    /// Get positions of all cursors sorted in reverse order (for editing)
-    #[allow(dead_code)]
-    pub fn positions_reversed(&self) -> Vec<CursorPos> {
-        let mut positions: Vec<CursorPos> = self.cursors.iter().map(|c| c.pos).collect();
-        positions.sort_by(|a, b| b.cmp(a)); // Reverse order
-        positions
-    }
-}
-
-impl Default for CursorSet {
-    fn default() -> Self {
-        Self::new(CursorPos::new(0, 0))
-    }
-}
-
-// ============================================================================
-// Selection
-// ============================================================================
-
-/// A text selection range
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Selection {
-    /// Start position (line, column)
-    pub start: (usize, usize),
-    /// End position (line, column)
-    pub end: (usize, usize),
-}
-
-impl Selection {
-    /// Create a new selection
-    pub fn new(start: (usize, usize), end: (usize, usize)) -> Self {
-        Self { start, end }
-    }
-
-    /// Get the normalized selection (start before end)
-    pub fn normalized(&self) -> Self {
-        if self.start.0 > self.end.0 || (self.start.0 == self.end.0 && self.start.1 > self.end.1) {
-            Self {
-                start: self.end,
-                end: self.start,
-            }
-        } else {
-            *self
-        }
-    }
-
-    /// Check if a position is within the selection
-    pub fn contains(&self, line: usize, col: usize) -> bool {
-        let norm = self.normalized();
-        if line < norm.start.0 || line > norm.end.0 {
-            return false;
-        }
-        if line == norm.start.0 && line == norm.end.0 {
-            col >= norm.start.1 && col < norm.end.1
-        } else if line == norm.start.0 {
-            col >= norm.start.1
-        } else if line == norm.end.0 {
-            col < norm.end.1
-        } else {
-            true
-        }
-    }
-}
-
-// ============================================================================
-// Find/Replace Types
-// ============================================================================
-
-/// Find/Replace mode
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
-pub enum FindReplaceMode {
-    #[default]
-    Find,
-    Replace,
-}
-
-/// Options for find/replace operations
-#[derive(Clone, Debug, Default)]
-pub struct FindOptions {
-    /// Case-sensitive search
-    pub case_sensitive: bool,
-    /// Match whole words only
-    pub whole_word: bool,
-    /// Use regex pattern
-    pub use_regex: bool,
-}
-
-/// A match found in the text
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct FindMatch {
-    /// Start position
-    pub start: CursorPos,
-    /// End position
-    pub end: CursorPos,
-}
-
-impl FindMatch {
-    /// Create a new find match
-    pub fn new(start: CursorPos, end: CursorPos) -> Self {
-        Self { start, end }
-    }
-}
-
-/// Find/Replace state
-#[derive(Clone, Debug, Default)]
-pub struct FindReplaceState {
-    /// Search query
-    pub query: String,
-    /// Replacement text
-    pub replace_with: String,
-    /// Search options
-    pub options: FindOptions,
-    /// All matches in document
-    pub matches: Vec<FindMatch>,
-    /// Currently focused match index
-    pub current_match: Option<usize>,
-    /// UI mode (Find or Replace)
-    pub mode: FindReplaceMode,
-    /// Input focus: true = query input, false = replace input
-    pub query_focused: bool,
-}
-
-impl FindReplaceState {
-    /// Create a new find/replace state
-    pub fn new(mode: FindReplaceMode) -> Self {
-        Self {
-            mode,
-            query_focused: true,
-            ..Default::default()
-        }
-    }
-
-    /// Get match count
-    pub fn match_count(&self) -> usize {
-        self.matches.len()
-    }
-
-    /// Get current match (1-indexed for display)
-    pub fn current_match_display(&self) -> usize {
-        self.current_match.map(|i| i + 1).unwrap_or(0)
-    }
-}
-
-// ============================================================================
-// Edit Operations
-// ============================================================================
-
-/// An edit operation for undo/redo
-#[derive(Clone, Debug)]
-enum EditOperation {
-    /// Insert text at position
-    Insert {
-        line: usize,
-        col: usize,
-        text: String,
-    },
-    /// Delete text at position
-    Delete {
-        line: usize,
-        col: usize,
-        text: String,
-    },
-    /// Insert a new line
-    InsertLine { line: usize, content: String },
-    /// Delete a line
-    DeleteLine { line: usize, content: String },
-    /// Merge with previous line
-    MergeLines { line: usize, col: usize },
-    /// Split line at position
-    SplitLine { line: usize, col: usize },
-}
 
 /// A multi-line text editor widget
 ///
@@ -1994,433 +1647,4 @@ impl_props_builders!(TextArea);
 /// Create a new text area
 pub fn textarea() -> TextArea {
     TextArea::new()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::super::syntax::{Language, SyntaxTheme};
-    use super::*;
-    use crate::layout::Rect;
-    use crate::render::Buffer;
-
-    #[test]
-    fn test_textarea_new() {
-        let ta = TextArea::new();
-        assert_eq!(ta.lines.len(), 1);
-        assert_eq!(ta.lines[0], "");
-    }
-
-    #[test]
-    fn test_textarea_content() {
-        let ta = TextArea::new().content("Hello\nWorld");
-        assert_eq!(ta.lines.len(), 2);
-        assert_eq!(ta.lines[0], "Hello");
-        assert_eq!(ta.lines[1], "World");
-    }
-
-    #[test]
-    fn test_textarea_insert_char() {
-        let mut ta = TextArea::new();
-        ta.insert_char('H');
-        ta.insert_char('i');
-        assert_eq!(ta.get_content(), "Hi");
-        assert_eq!(ta.cursor_position(), (0, 2));
-    }
-
-    #[test]
-    fn test_textarea_insert_newline() {
-        let mut ta = TextArea::new().content("Hello");
-        ta.set_cursor(0, 5);
-        ta.insert_char('\n');
-        assert_eq!(ta.lines.len(), 2);
-        assert_eq!(ta.cursor_position(), (1, 0));
-    }
-
-    #[test]
-    fn test_textarea_delete_char_before() {
-        let mut ta = TextArea::new().content("Hello");
-        ta.set_cursor(0, 3);
-        ta.delete_char_before();
-        assert_eq!(ta.get_content(), "Helo");
-        assert_eq!(ta.cursor_position(), (0, 2));
-    }
-
-    #[test]
-    fn test_textarea_delete_char_at() {
-        let mut ta = TextArea::new().content("Hello");
-        ta.set_cursor(0, 2);
-        ta.delete_char_at();
-        assert_eq!(ta.get_content(), "Helo");
-    }
-
-    #[test]
-    fn test_textarea_movement() {
-        let mut ta = TextArea::new().content("Hello\nWorld");
-        ta.set_cursor(0, 2);
-
-        ta.move_right();
-        assert_eq!(ta.cursor_position(), (0, 3));
-
-        ta.move_left();
-        assert_eq!(ta.cursor_position(), (0, 2));
-
-        ta.move_down();
-        assert_eq!(ta.cursor_position(), (1, 2));
-
-        ta.move_up();
-        assert_eq!(ta.cursor_position(), (0, 2));
-
-        ta.move_home();
-        assert_eq!(ta.cursor_position(), (0, 0));
-
-        ta.move_end();
-        assert_eq!(ta.cursor_position(), (0, 5));
-    }
-
-    #[test]
-    fn test_textarea_undo_redo() {
-        let mut ta = TextArea::new();
-        ta.insert_char('A');
-        ta.insert_char('B');
-        assert_eq!(ta.get_content(), "AB");
-
-        ta.undo();
-        assert_eq!(ta.get_content(), "A");
-
-        ta.undo();
-        assert_eq!(ta.get_content(), "");
-
-        ta.redo();
-        assert_eq!(ta.get_content(), "A");
-
-        ta.redo();
-        assert_eq!(ta.get_content(), "AB");
-    }
-
-    #[test]
-    fn test_textarea_selection() {
-        let mut ta = TextArea::new().content("Hello World");
-        // Create selection from position 0 to 5
-        ta.set_cursor(0, 0);
-        ta.start_selection();
-        ta.set_cursor(0, 5);
-        ta.cursors.primary_mut().anchor = Some(CursorPos::new(0, 0));
-
-        let selected = ta.get_selection();
-        assert_eq!(selected, Some("Hello".to_string()));
-    }
-
-    #[test]
-    fn test_textarea_delete_selection() {
-        let mut ta = TextArea::new().content("Hello World");
-        // Create selection from position 0 to 6
-        ta.set_cursor(0, 6);
-        ta.cursors.primary_mut().anchor = Some(CursorPos::new(0, 0));
-        ta.delete_selection();
-        assert_eq!(ta.get_content(), "World");
-    }
-
-    #[test]
-    fn test_textarea_render() {
-        let ta = TextArea::new()
-            .content("Line 1\nLine 2\nLine 3")
-            .line_numbers(true);
-
-        let mut buffer = Buffer::new(40, 10);
-        let area = Rect::new(0, 0, 40, 10);
-        let mut ctx = RenderContext::new(&mut buffer, area);
-
-        ta.render(&mut ctx);
-    }
-
-    #[test]
-    fn test_textarea_word_movement() {
-        let mut ta = TextArea::new().content("Hello World Test");
-        ta.set_cursor(0, 0);
-
-        ta.move_word_right();
-        assert_eq!(ta.cursor_position().1, 6); // After "Hello "
-
-        ta.move_word_right();
-        assert_eq!(ta.cursor_position().1, 12); // After "World "
-
-        ta.move_word_left();
-        assert_eq!(ta.cursor_position().1, 6);
-    }
-
-    #[test]
-    fn test_textarea_handle_key() {
-        let mut ta = TextArea::new();
-
-        assert!(ta.handle_key(&Key::Char('A')));
-        assert_eq!(ta.get_content(), "A");
-
-        assert!(ta.handle_key(&Key::Enter));
-        assert_eq!(ta.lines.len(), 2);
-
-        assert!(ta.handle_key(&Key::Backspace));
-        assert_eq!(ta.lines.len(), 1);
-    }
-
-    #[test]
-    fn test_textarea_delete_line() {
-        let mut ta = TextArea::new().content("Line 1\nLine 2\nLine 3");
-        ta.set_cursor(1, 0);
-        ta.delete_line();
-        assert_eq!(ta.line_count(), 2);
-        assert_eq!(ta.get_content(), "Line 1\nLine 3");
-    }
-
-    #[test]
-    fn test_textarea_duplicate_line() {
-        let mut ta = TextArea::new().content("Line 1\nLine 2");
-        ta.set_cursor(0, 0);
-        ta.duplicate_line();
-        assert_eq!(ta.line_count(), 3);
-        assert_eq!(ta.get_content(), "Line 1\nLine 1\nLine 2");
-    }
-
-    #[test]
-    fn test_textarea_select_all() {
-        let mut ta = TextArea::new().content("Hello\nWorld");
-        ta.select_all();
-        assert!(ta.has_selection());
-        let sel = ta.cursors.primary().selection().unwrap();
-        assert_eq!(sel.start, (0, 0));
-        assert_eq!(sel.end, (1, 5));
-    }
-
-    #[test]
-    fn test_selection_contains() {
-        let sel = Selection::new((0, 5), (0, 10));
-        assert!(!sel.contains(0, 4));
-        assert!(sel.contains(0, 5));
-        assert!(sel.contains(0, 7));
-        assert!(!sel.contains(0, 10));
-    }
-
-    #[test]
-    fn test_textarea_helper() {
-        let ta = textarea();
-        assert_eq!(ta.line_count(), 1);
-    }
-
-    #[test]
-    fn test_textarea_read_only() {
-        let mut ta = TextArea::new().read_only(true);
-        ta.insert_char('A');
-        assert_eq!(ta.get_content(), "");
-    }
-
-    #[test]
-    fn test_textarea_max_lines() {
-        let mut ta = TextArea::new().max_lines(2);
-        ta.insert_char('A');
-        ta.insert_char('\n');
-        ta.insert_char('B');
-        ta.insert_char('\n'); // Should not add third line
-        ta.insert_char('C');
-        assert_eq!(ta.line_count(), 2);
-    }
-
-    #[test]
-    fn test_textarea_syntax_highlighting() {
-        let ta = TextArea::new()
-            .content("fn main() {\n    println!(\"Hello\");\n}")
-            .syntax(Language::Rust);
-
-        assert_eq!(ta.get_syntax_language(), Language::Rust);
-    }
-
-    #[test]
-    fn test_textarea_syntax_with_theme() {
-        let ta = TextArea::new()
-            .content("def hello():\n    print('Hi')")
-            .syntax_with_theme(Language::Python, SyntaxTheme::monokai());
-
-        assert_eq!(ta.get_syntax_language(), Language::Python);
-    }
-
-    #[test]
-    fn test_textarea_set_language() {
-        let mut ta = TextArea::new().content("console.log('Hello');");
-
-        ta.set_language(Language::JavaScript);
-        assert_eq!(ta.get_syntax_language(), Language::JavaScript);
-
-        ta.set_language(Language::None);
-        assert_eq!(ta.get_syntax_language(), Language::None);
-    }
-
-    #[test]
-    fn test_textarea_syntax_render() {
-        let ta = TextArea::new()
-            .content("fn test() {}")
-            .syntax(Language::Rust);
-
-        let mut buffer = Buffer::new(40, 5);
-        let area = Rect::new(0, 0, 40, 5);
-        let mut ctx = RenderContext::new(&mut buffer, area);
-
-        // Should not panic when rendering with syntax highlighting
-        ta.render(&mut ctx);
-    }
-
-    // =========================================================================
-    // Find/Replace Tests
-    // =========================================================================
-
-    #[test]
-    fn test_find_basic() {
-        let mut ta = TextArea::new().content("Hello World");
-        ta.open_find();
-        ta.set_find_query("World");
-
-        let state = ta.find_state().unwrap();
-        assert_eq!(state.matches.len(), 1);
-        assert_eq!(state.matches[0].start, CursorPos::new(0, 6));
-        assert_eq!(state.matches[0].end, CursorPos::new(0, 11));
-    }
-
-    #[test]
-    fn test_find_case_insensitive() {
-        let mut ta = TextArea::new().content("Hello hello HELLO");
-        ta.open_find();
-        ta.set_find_query("hello");
-
-        // Default is case insensitive
-        let state = ta.find_state().unwrap();
-        assert_eq!(state.matches.len(), 3);
-    }
-
-    #[test]
-    fn test_find_case_sensitive() {
-        let mut ta = TextArea::new().content("Hello hello HELLO");
-        ta.open_find();
-        ta.toggle_case_sensitive();
-        ta.set_find_query("hello");
-
-        let state = ta.find_state().unwrap();
-        assert_eq!(state.matches.len(), 1);
-    }
-
-    #[test]
-    fn test_find_whole_word() {
-        let mut ta = TextArea::new().content("test testing tested");
-        ta.open_find();
-        ta.toggle_whole_word();
-        ta.set_find_query("test");
-
-        let state = ta.find_state().unwrap();
-        assert_eq!(state.matches.len(), 1);
-        assert_eq!(state.matches[0].start.col, 0);
-    }
-
-    #[test]
-    fn test_find_next_wraps() {
-        let mut ta = TextArea::new().content("aaa");
-        ta.open_find();
-        ta.set_find_query("a");
-
-        assert_eq!(ta.find_state().unwrap().current_match, Some(0));
-        ta.find_next();
-        assert_eq!(ta.find_state().unwrap().current_match, Some(1));
-        ta.find_next();
-        assert_eq!(ta.find_state().unwrap().current_match, Some(2));
-        ta.find_next();
-        assert_eq!(ta.find_state().unwrap().current_match, Some(0)); // Wrapped
-    }
-
-    #[test]
-    fn test_replace_current() {
-        let mut ta = TextArea::new().content("foo bar foo");
-        ta.open_replace();
-        ta.set_find_query("foo");
-        ta.set_replace_text("baz");
-
-        ta.replace_current();
-        assert_eq!(ta.get_content(), "baz bar foo");
-    }
-
-    #[test]
-    fn test_replace_all() {
-        let mut ta = TextArea::new().content("foo bar foo");
-        ta.open_replace();
-        ta.set_find_query("foo");
-        ta.set_replace_text("baz");
-
-        ta.replace_all();
-        assert_eq!(ta.get_content(), "baz bar baz");
-    }
-
-    #[test]
-    fn test_close_find() {
-        let mut ta = TextArea::new().content("test");
-        ta.open_find();
-        assert!(ta.is_find_open());
-        ta.close_find();
-        assert!(!ta.is_find_open());
-    }
-
-    // =========================================================================
-    // Multiple Cursors Tests
-    // =========================================================================
-
-    #[test]
-    fn test_add_cursor() {
-        let mut ta = TextArea::new().content("Hello\nWorld\nTest");
-        ta.add_cursor_at(1, 2);
-        assert_eq!(ta.cursor_count(), 2);
-    }
-
-    #[test]
-    fn test_add_cursor_above() {
-        let mut ta = TextArea::new().content("Line1\nLine2\nLine3");
-        ta.set_cursor(1, 3);
-        ta.add_cursor_above();
-        assert_eq!(ta.cursor_count(), 2);
-        let positions = ta.cursor_positions();
-        assert!(positions.contains(&(0, 3)));
-        assert!(positions.contains(&(1, 3)));
-    }
-
-    #[test]
-    fn test_add_cursor_below() {
-        let mut ta = TextArea::new().content("Line1\nLine2\nLine3");
-        ta.set_cursor(1, 3);
-        ta.add_cursor_below();
-        assert_eq!(ta.cursor_count(), 2);
-        let positions = ta.cursor_positions();
-        assert!(positions.contains(&(1, 3)));
-        assert!(positions.contains(&(2, 3)));
-    }
-
-    #[test]
-    fn test_clear_secondary_cursors() {
-        let mut ta = TextArea::new().content("Hello\nWorld");
-        ta.add_cursor_at(1, 0);
-        assert_eq!(ta.cursor_count(), 2);
-        ta.clear_secondary_cursors();
-        assert_eq!(ta.cursor_count(), 1);
-    }
-
-    #[test]
-    fn test_select_next_occurrence() {
-        let mut ta = TextArea::new().content("foo bar foo baz foo");
-        ta.set_cursor(0, 0);
-
-        // Select "foo" (set cursor at position and give it a word)
-        ta.select_next_occurrence();
-        // First occurrence at 0,0 should now have a cursor
-        assert!(ta.cursor_count() >= 1);
-    }
-
-    #[test]
-    fn test_cursor_positions() {
-        let mut ta = TextArea::new().content("Hello\nWorld");
-        ta.add_cursor_at(1, 2);
-
-        let positions = ta.cursor_positions();
-        assert_eq!(positions.len(), 2);
-    }
 }
